# Speed up composition