--- a/unicode-data/stack.yaml
+++ b/unicode-data/stack.yaml
@@ -1,5 +1,3 @@
-# This stack.yaml is to independently build
-# the utility using its own cabal file
-resolver: lts-11.0
-extra-deps:
-- getopt-generics-0.13.0.2
+packages:
+- '.'
+resolver: lts-15.10