… Data/Unicode/Properties/CombiningClass.hs
… Data/Unicode/Properties/Compositions.hs
… Data/Unicode/Properties/Decomposable.hs
… Data/Unicode/Properties/DecomposableK.hs
… Data/Unicode/Properties/DecomposeHangul.hs
… unicode-data/UCD2Haskell.hs