I spent some time refactoring `composeChar`. Results are as follows:

```
Benchmark         Before     After
------------------------------------
NFC/Japanese     72.41 ms   35.08 ms
NFC/Vietnamese   60.76 ms   32.88 ms
NFC/AllChars     54.81 ms   33.42 ms
NFC/Deutsch      54.64 ms   25.72 ms
NFC/English      53.56 ms   24.60 ms
NFC/Devanagari   54.19 ms   29.24 ms
NFC/Korean       52.33 ms   33.94 ms
NFKC/Japanese    72.44 ms   38.35 ms
NFKC/Vietnamese  60.08 ms   34.33 ms
NFKC/AllChars    68.44 ms   43.96 ms
NFKC/Deutsch     57.80 ms   28.27 ms
NFKC/English     54.64 ms   26.53 ms
NFKC/Devanagari  54.30 ms   30.35 ms
NFKC/Korean      53.73 ms   34.38 ms
```

During refactoring I noticed that the trick with `FalseA | FalseB | FalseC | TrueA` does not seem to be needed anymore and bitmaps inline well enough, so I took a liberty to remove it.

A new bitmap `composePairSecondNonCombining` lists non-combining characters, which can appear in the second half of `composePair`. All such compositions are listed in `composePairNonCombining`. This is a fairly sparse bitmap with only 40 non-zero bits, but I was not able to come up with any simpler representation without performance degradation.

With regards to `composeChar` I followed suggestions to straighten up recursive calls between `composeChar` / `decomposeAll` and to use an `Either`-like structure to switch between mutually exclusive `(Maybe Char, ReBuf)` and `JamoBuf`.