@harendra-kumar 2020-05-04
I was wondering if the FalseA/FalseB thing still helps. Good riddance. 

The overall code looks much better now. Thank you! I have a couple of formatting changes and a perf improvement on top of your changes. I am going to merge this and create another PR for those changes. I am wondering if we can we write `decomposeChar` as well in the same style as `composeChar`? And use SPEC constr optimization on the recursive function in the same way as I used in `composeChar` (in the upcoming PR).